Move theano.tensor/scalar utility functions to more appropriate modules
